refactor: remove AuthSpec from WebSpec
